honor sr subject, update claudemd, update changelog
